client review section

+++ src/Dashboard/AddPackages/RemovePackages/RemovePackages.js
@@ -1,23 +1,25 @@
 import React, { useEffect, useState } from 'react';
+import Swal from 'sweetalert2';
 
 const RemovePackages = () => {
     const [packages, setPackages] = useState([]);
-    const [control, setControl] = useState(true);
+    const [isDeleted, setIsDeleted] = useState(true);
     const [isLoading, setIsLoading] = useState(true);
 
 
     useEffect(() => {
         setIsLoading(true)
-        fetch("https://fierce-falls-08266.herokuapp.com/packages")
+        fetch("http://localhost:5000/packages")
           .then((res) => res.json())
           .then((data) => {
             setPackages(data)
+            console.log(data);
             setIsLoading(false)
           })
-      }, [control]);
+      }, [isDeleted]);
 
     //   const handleDelete = (id) =>{
-    //     axios.delete(`https://fierce-falls-08266.herokuapp.com/packages/${id}`)
+    //     axios.delete(`http://localhost:5000/packages/${id}`)
     //     .then(res => res.json())
     //      .then((data) => {
     //             if (data.deletedCount) {
@@ -25,21 +27,65 @@ const RemovePackages = () => {
     //             }
     //       })}
 
-        const handleDelete = id => {
-            setIsLoading(true)
-            const url = `https://fierce-falls-08266.herokuapp.com/packages/${id}`;
-            fetch(url, {
-            method: 'DELETE'
-            })
-            .then(res => res.json())
-            .then(data => {
-            if (data.deletedCount) {
-            setControl(!control);
-            setIsLoading(false)
+        // const handleDelete = id => {
+        //     setIsLoading(true)
+        //     const url = `http://localhost:5000/packages/${id}`;
+        //     fetch(url, {
+        //     method: 'DELETE'
+        //     })
+        //     .then(res => res.json())
+        //     .then(data => {
+        //     if (data.deletedCount) {
+        //     setControl(!control);
+        //     setIsLoading(false)
 
-            }
+        //     }
+        //     })
+        //     }
+            // delete a order
+    const handleDelete = (id) => {
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You Want to Delete it!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!'
+      })
+  
+      .then((result) => {
+        if (result.isConfirmed) {
+          fetch(`http://localhost:5000/packages/${id}`, {
+            method: "DELETE",
+          })
+            .then((res) => res.json())
+            .then((data) => {
+              if (data) {
+                Swal.fire(
+                  "Deleted!",
+                  "Your file has been deleted.",
+                  "success"
+                );
+                setIsDeleted(data);
+                window.location.reload();
+              }
             })
-            }
+            .finally(() => {
+              setIsDeleted(false);
+            });
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+            Swal.fire(
+            "Cancelled",
+            "Your imaginary file is safe :)",
+            "error"
+          );
+        }
+      });
+  };
 
     return (
         <div className='remove-packages-section'>
@@ -52,44 +98,24 @@ const RemovePackages = () => {
           <thead>
             <tr>
               <th scope="col">#</th>
-              <th>Proparty Name</th>
-              <th>User Email</th>
-              <th>Address</th>
-              <th>Status</th>
+              <th>Place</th>
+              <th>image</th>
+              <th>Price</th>
+              <th>time</th>
               <th>action</th>
             </tr>
           </thead>
           {
             packages?.map((pd, index)=>(
-<tbody key={index}>
+      <tbody key={index}>
         
-            <tr>
-              <th scope="row">1</th>
-              <td>{pd?.data?.fullName}</td>
-              <td>{pd?.data?.email}</td>
-              <td>{pd?.address}</td>
-              <td>{pd?.data?.status}</td>
-              <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">
-        Launch demo modal
-            </button>
-            <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="exampleModalLabel">Modal title</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        ...
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-primary" onClick={() => handleDelete(pd?._id)}>Delete</button>
-        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancle</button>
-      </div>
-    </div>
-  </div>
-</div>
-            </tr>
+      <tr>
+        <th scope="row">{index + 1}</th>
+        <td>{pd?.place_name} <span>{pd?.Country}</span></td>
+        <td><img class="img-thumbnail" src={pd?.image} alt="" /></td>
+        <td>{pd?.price}</td>
+        <td>{pd?.time}</td>
+        <button type="button" class="btn btn-danger bg-danger" onClick={() => handleDelete(pd?._id)}>Delete</button>            </tr>
           </tbody>
     ))
 }
